client: update time stamp for non-changed index
When an index is stale but there are no changes to it at the server
side, update the retrieved timestamp for the index. Otherwise, the
server is asked for index change on each future invocation, which
cancels the benefits of a cache. Further, index retrieval is called on
each client.Get, so loading the vulnerabilities can be slowed down too.

Change-Id: Ie4814d5f74b412e746247097939c3c09cbbb87f5
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/351869
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Roland Shoemaker <roland@golang.org>
Trust: Julie Qiu <julie@golang.org>