cmd/govulncheck: show rel path for call stacks if shorter than abs
In theory, govulncheck can be called on a package that is not present in
current dir, but is present in std lib or mod cache. In that case, the
relative path to files in such packages could be longer and more
confusing than the absolute path.

Change-Id: Iaa14caad4756adc924fbe21fb713a053f9ccd8b2
Reviewed-on: https://go-review.googlesource.com/c/vuln/+/427337
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>